backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.26.0
google-auth==2.48.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
google-genai==1.64.0
slack-sdk==3.26.2
spacy==3.7.2
transformers==4.37.2
torch==2.1.2
sentence-transformers==2.3.1
numpy==1.26.3
pandas==2.2.0
python-docx==1.1.0
PyPDF2==3.0.1
markdown==3.5.2
reportlab==4.0.9
openpyxl==3.1.2
bcrypt==4.0.1
psycopg2-binary==2.9.9
openai==1.3.0
scikit-learn==1.3.2
